wip: Wave A — dialect kernel, testing database seam, store split
The four Wave A buckets, gated. Delivered:

- src/sql/Dialect.ts and src/sql/Mutations.ts: the dialect kernel (dialectOf,
  booleanCodec, lockClause, identifier, columnType) and the six mutation
  helpers, with the mysql branches proven as rendered SQL text. Nothing
  imports them yet.
- src/testing/Database.ts and the sqlite/postgres/mysql subpaths: the
  TestDatabase seam, four providers, the PgliteClient leak removed from every
  exported testing layer type. vitest isolate:false, the four scripts, Node 22,
  five parallel CI jobs.
- test/sql/** and test/fields/**: the store, migration and custom-field
  contracts made dialect-neutral, plus Errors, Concurrency and Dialect.
- src/sql/stores/**: SqlStores.ts split six ways as a pure move; the .d.ts is
  unchanged but for two import-type modifiers.

Gate fixes: the pg-only `count(*)::int` in test/email-otp; SQLite reports a
PRIMARY KEY conflict as 1555, not 2067, so the phone number is a NOT NULL
UNIQUE column there and a second claim classifies as a lost race again;
`sign_count` normalised from the string PostgreSQL's driver hands back for a
bigint, as `last_used_step` already is.

Gates: check zero lines; pglite 1482/1482; pg 1482/1482; sqlite 14 failures in
four plugin catalog-shape tests (B6 ports them); mysql 61 files on the
unported core migrations (B1). WIP because sqlite is not yet green.

+++ test/email-otp/EmailOtp.test.ts
@@ -25,7 +25,7 @@ const codeFor = Effect.fnUntraced(function* (
 const rowsAt = Effect.fnUntraced(function* (identifier: string) {
   const sql = yield* SqlClient.SqlClient
   const rows = yield* sql<{ readonly count: number | string }>`
-    select count(*)::int as count from verifications where identifier = ${identifier}`
+    select count(*) as count from verifications where identifier = ${identifier}`
   return Number(rows[0]?.count ?? 0)
 })
